+++ docs-src/vite.config.js
@@ -4,4 +4,9 @@ import vue from '@vitejs/plugin-vue'
 export default defineConfig({
   base: '/vue-sanitize-directive/',
   plugins: [vue()],
+  server: {
+    fs: {
+      allow: ['..'],
+    },
+  },
 })
