+++ package.json
@@ -12,23 +12,35 @@
     "sanitize",
     "v-sanitize"
   ],
-  "main": "dist/vue-sanitize-directive.umd.js",
-  "module": "dist/vue-sanitize-directive.esm.js",
+  "type": "module",
   "exports": {
     ".": {
-      "import": "./dist/vue-sanitize-directive.esm.js",
-      "require": "./dist/vue-sanitize-directive.umd.js"
+      "import": "./dist/vue3/vue-sanitize-directive.js",
+      "require": "./dist/vue3/vue-sanitize-directive.cjs"
+    },
+    "./vue2": {
+      "import": "./dist/vue2/vue-sanitize-directive.js",
+      "require": "./dist/vue2/vue-sanitize-directive.cjs"
+    },
+    "./vue3": {
+      "import": "./dist/vue3/vue-sanitize-directive.js",
+      "require": "./dist/vue3/vue-sanitize-directive.cjs"
     }
   },
   "scripts": {
-    "build": "vite build"
+    "build": "tsup"
   },
   "license": "MIT",
   "dependencies": {
     "sanitize-html": "^2.7.1"
   },
   "devDependencies": {
-    "vite": "^3.0.4"
+    "@types/sanitize-html": "^2.6.2",
+    "tsup": "^6.2.2",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.9",
+    "vue2": "npm:vue@^2.7.9",
+    "vue3": "npm:vue@^3.2.37"
   },
   "directories": {
     "doc": "docs"
