+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "target": "ES2017",
+    "module": "commonjs",
+    "lib": ["DOM", "ES2017"],
+    "strict": true,
+    "esModuleInterop": true
+  }
+}
